Rename streaming.py to dataloader.py (per review)
Move the DataLoader, IterableDataset, shuffle_buffer, and parquet_source
into pymc_extras/variational/dataloader.py, with matching test files.
Pure rename; no behavior change.
