+++ pymc_extras/variational/__init__.py
@@ -11,7 +11,7 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
-from pymc_extras.variational.streaming import (
+from pymc_extras.variational.dataloader import (
     DataLoader,
     IterableDataset,
     parquet_source,
