+++ pymc_extras/variational/dataloader.py
